feat: moves form admin to separate class.
